requirements.txt
beautifulsoup4==4.10.0
certifi==2023.7.22
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.5
cssselect==1.1.0
dataclasses==0.8
EbookLib==0.17.1
Flask==2.0.1
idna==3.4
importlib-metadata==4.8.3
install==1.3.5
itsdangerous==2.0.1
Jinja2==3.0.3
lxml==4.9.3
MarkupSafe==2.0.1
Pillow==8.3.1
requests==2.26.0
six==1.16.0
soupsieve==2.3.2.post1
typing_extensions==4.1.1
urllib3==1.26.16
Werkzeug==2.0.3
zipp==3.6.0
